requirements.txt
flake8==6.0.0
autopep8==2.0.1
pylint==2.16.4

Django==4.2
tweepy==4.12.1

pandas==1.4.4
pandas-profiling==3.6.6

python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
joblib==1.2.0

scikit-learn==1.2.1
spacy==3.5.0
tqdm==4.64.1
pyspellchecker==0.7.1